Experimenting with shipyard app. This will be fleshed out.

+++ crates/mani/fabled_render/src/mesh/container/mesh_data.rs
@@ -1,10 +1,9 @@
 use crate::mesh::container::Vertex;
 use crate::mesh::Indices;
 
-#[derive(Debug)]
-#[repr(C, align(16))]
+#[derive(Debug, Clone)]
 pub struct Mesh {
     pub vertices: Vec<Vertex>,
     pub indices: Indices,
-    pub material_id: u32,
+    pub material_id: usize,
 }
